refactor : attraction 레포지토리 의존성 제거
코드의 복잡성과 간결성을 위해 변경
