+++ src/main/java/com/rualone/app/domain/attraction/dao/AttractionInfoRepository.java
@@ -3,5 +3,8 @@ package com.rualone.app.domain.attraction.dao;
 import com.rualone.app.domain.attraction.entity.AttractionInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttractionInfoRepository extends JpaRepository<AttractionInfo, Integer>, AttractionInfoRepositoryCustom {
+import java.util.Optional;
+
+public interface AttractionInfoRepository extends JpaRepository<AttractionInfo, Integer>{
+    Optional<AttractionInfo> findByContentId(Integer contentId);
 }
